add --resize option and combine _full_res with _low_res scripts into one unifiled one
